ROM-114 date系のutil関数を実装する。
